Apparently, the f_contexts in the boost.context library DO NOT maintain the signal mask on context switches...

+++ controller/src/controller/resources/run/run.cpp
@@ -73,9 +73,25 @@ namespace run{
         } else {
             ctx_ptr = std::make_shared<controller::app::ExecutionContext>(controller::app::ExecutionContext::run);
         }
+        sigset_t sigmask = {};
+        int status = sigemptyset(&sigmask);
+        if(status == -1){
+            std::cerr << "run.cpp:79:sigemptyset failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+            throw "what?";
+        }
+        status = sigaddset(&sigmask, SIGCHLD);
+        if(status == -1){
+            std::cerr << "run.cpp:84:sigaddmask failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+            throw "what?";
+        }
         for (auto& relation: ctx_ptr->manifest()){
             boost::context::fiber f{
-                [&, req, ctx_ptr, relation](boost::context::fiber&& g) {
+                [&, req, ctx_ptr, relation, sigmask](boost::context::fiber&& g) {
+                    int status = sigprocmask(SIG_BLOCK, &sigmask, nullptr);
+                    if(status == -1){
+                        std::cerr << "run.cpp:92:sigprocmask failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                        throw "what?";
+                    }
                      // Get the index of the relation.
                     auto it = std::find(ctx_ptr->manifest().begin(), ctx_ptr->manifest().end(), relation);
                     std::ptrdiff_t idx = it - ctx_ptr->manifest().begin();
@@ -124,26 +140,26 @@ namespace run{
                         case 0:
                         {
                             //Child Process.
-                            sigset_t sigmask = {};
-                            int status = sigemptyset(&sigmask);
-                            if(status == -1){
-                                std::cerr << "run.cpp:130:sigemptyset failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
-                                throw "what?";
-                            }
-                            status = sigaddset(&sigmask, SIGCHLD);
-                            if(status == -1){
-                                std::cerr << "run.cpp:135:sigaddmask failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
-                                throw "what?";
-                            }
-                            status = sigprocmask(SIG_BLOCK, &sigmask, nullptr);
-                            if(status == -1){
-                                std::cerr << "run.cpp:140:sigprocmask failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
-                                throw "what?";
-                            }
+                            // sigset_t sigmask = {};
+                            // int status = sigemptyset(&sigmask);
+                            // if(status == -1){
+                            //     std::cerr << "run.cpp:146:sigemptyset failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                            //     throw "what?";
+                            // }
+                            // status = sigaddset(&sigmask, SIGCHLD);
+                            // if(status == -1){
+                            //     std::cerr << "run.cpp:151:sigaddmask failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                            //     throw "what?";
+                            // }
+                            // status = sigprocmask(SIG_BLOCK, &sigmask, nullptr);
+                            // if(status == -1){
+                            //     std::cerr << "run.cpp:156:sigprocmask failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                            //     throw "what?";
+                            // }
                             errno = 0;
                             status = nice(2);
                             if(status == -1 && errno != 0){
-                                std::cerr << "run.cpp:146:nice failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                std::cerr << "run.cpp:162:nice failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                             }
                             setpgid(0,0);
                             std::uint64_t notice = 1;
@@ -161,7 +177,7 @@ namespace run{
                                         case EINTR:
                                             break;
                                         default:
-                                            std::cerr << "run.cpp:164:poll() failed with error:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                            std::cerr << "run.cpp:1804:poll() failed with error:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                                             throw "what?";
                                     }
                                 } else if(pfds[0].revents & POLLOUT){
@@ -174,17 +190,17 @@ namespace run{
                                             case EWOULDBLOCK:
                                                 break;
                                             default:
-                                                std::cerr << "run.cpp:177:write() failed with error:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                                std::cerr << "run.cpp:193:write() failed with error:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                                                 throw "what?";
                                         }
                                     }
                                 } else {
-                                    std::cerr << "run.cpp:182:poll() has an unrecognized flag." << std::endl;
+                                    std::cerr << "run.cpp:198:poll() has an unrecognized flag." << std::endl;
                                     throw "what?";
                                 }
                             } while(errno == EINTR || errno == EWOULDBLOCK);
                             if(close(sync_fd) == -1){
-                                std::cerr << "run.cpp:187:closing sync_fd in the child failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                std::cerr << "run.cpp:203:closing sync_fd in the child failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                                 throw "what?";
                             }
                             if( close(downstream[1]) == -1 ){
@@ -234,12 +250,11 @@ namespace run{
                         0
                     };
                     // Block until the child process sets the pgid.
-                    int status;
                     int len;
                     do{
                         status = poll(&pfds[0], 1, -1);
                         if(status == -1){
-                            std::cerr << "run.cpp:242:poll() failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                            std::cerr << "run.cpp:257:poll() failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                             switch(errno)
                             {
                                 case EINTR:
@@ -251,7 +266,7 @@ namespace run{
                             std::uint64_t notice = 0;
                             len = read(sync_fd, &notice, sizeof(notice));
                             if(len == -1){
-                                std::cerr << "run.cpp:254:read() failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                std::cerr << "run.cpp:269:read() failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                                 switch(errno)
                                 {
                                     case EINTR:
@@ -263,7 +278,7 @@ namespace run{
                                 }
                             }
                         } else {
-                            std::cerr << "run.cpp:266:unrecognized events on poll" << std::endl;
+                            std::cerr << "run.cpp:281:unrecognized events on poll" << std::endl;
                             throw "what?";
                         }
                     }while(errno == EINTR || errno == EWOULDBLOCK);                    
@@ -271,7 +286,7 @@ namespace run{
                     ctx_ptr->thread_controls()[idx].pid() = pid;    
                     ctx_ptr->synchronize();
                     if(close(sync_fd) == -1){
-                        std::cerr << "run.cpp:274:Closing the write side of sync_fd in the parent process failed: " << std::make_error_code(std::errc(errno)).message() << std::endl;
+                        std::cerr << "run.cpp:289:Closing the write side of sync_fd in the parent process failed: " << std::make_error_code(std::errc(errno)).message() << std::endl;
                         throw "This shouldn't happen.";
                     }
                     //Parent Process.
@@ -295,7 +310,7 @@ namespace run{
                                 case EINTR:
                                     break;
                                 default:
-                                    std::cerr << "run.cpp:298:poll failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                    std::cerr << "run.cpp:313:poll failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                                     throw "what?";
                             }
                         } else if (pfds[1].revents & POLLIN || pfds[1].revents & POLLHUP){
@@ -307,12 +322,12 @@ namespace run{
                                     case EINTR:
                                         break;
                                     default:
-                                        std::cerr << "run.cpp:310:ready failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                        std::cerr << "run.cpp:325:ready failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                                         throw "what?";
                                 }
                             }
                         } else {
-                            std::cerr << "run.cpp:315:pfds has no recognized events." << std::endl;
+                            std::cerr << "run.cpp:330:pfds has no recognized events." << std::endl;
                             throw "what?";
                         }
                     }while(errno == EINTR);
@@ -321,10 +336,10 @@ namespace run{
                         struct timespec ts = {};
                         int status = clock_gettime(CLOCK_REALTIME, &ts);
                         if(status == -1){
-                            std::cerr << "run.cpp:324:clock_gettime failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
-                            std::cerr << "run.cpp:325:pausing the action launcher failed:" << std::make_error_code(std::errc(errsv)).message() << std::endl;
+                            std::cerr << "run.cpp:339:clock_gettime failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                            std::cerr << "run.cpp:340:pausing the action launcher failed:" << std::make_error_code(std::errc(errsv)).message() << std::endl;
                         } else {
-                            std::cerr << "run.cpp:327:" << (ts.tv_sec*1000 + ts.tv_nsec/1000000) << ":pausing the action launcher failed:" << std::make_error_code(std::errc(errsv)).message() << std::endl;
+                            std::cerr << "run.cpp:342:" << (ts.tv_sec*1000 + ts.tv_nsec/1000000) << ":pausing the action launcher failed:" << std::make_error_code(std::errc(errsv)).message() << std::endl;
                         }
                         switch(errsv)
                         {
@@ -336,16 +351,21 @@ namespace run{
                     }
 
                     g = std::move(g).resume();
+                    status = sigprocmask(SIG_BLOCK, &sigmask, nullptr);
+                    if(status == -1){
+                        std::cerr << "controller-app.cpp:356:sigprocmask failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                        throw "what?";
+                    }
 
                     if (kill(-pid, SIGCONT) == -1){
                         int errsv = errno;
                         struct timespec ts = {};
                         int status = clock_gettime(CLOCK_REALTIME, &ts);
                         if(status == -1){
-                            std::cerr << "run.cpp:345:clock_gettime failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
-                            std::cerr << "run.cpp:346:unpausing the action launcher failed:" << std::make_error_code(std::errc(errsv)).message() << std::endl;
+                            std::cerr << "run.cpp:365:clock_gettime failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                            std::cerr << "run.cpp:366:unpausing the action launcher failed:" << std::make_error_code(std::errc(errsv)).message() << std::endl;
                         } else {
-                            std::cerr << "run.cpp:348:" << (ts.tv_sec*1000 + ts.tv_nsec/1000000) << ":unpausing the action launcher failed:" << std::make_error_code(std::errc(errsv)).message() << std::endl;
+                            std::cerr << "run.cpp:368:" << (ts.tv_sec*1000 + ts.tv_nsec/1000000) << ":unpausing the action launcher failed:" << std::make_error_code(std::errc(errsv)).message() << std::endl;
                         }
                         switch(errsv)
                         {
@@ -384,11 +404,11 @@ namespace run{
                                 case EINTR:
                                     break;
                                 default:
-                                    std::cerr << "run.cpp:387:poll failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                    std::cerr << "run.cpp:407:poll failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                                     throw "what?";
                             }
                         } else if(pfds[2].revents & POLLERR) {
-                            std::cerr << "run.cpp:391:the child process has closed the read end of the downstream pipe." << std::endl;
+                            std::cerr << "run.cpp:411:the child process has closed the read end of the downstream pipe." << std::endl;
                             throw "what?";
                         } else if(pfds[2].revents & POLLOUT){
                             len = write(downstream[1], params.data(), params.size());
@@ -398,7 +418,7 @@ namespace run{
                                     case EINTR:
                                         break;
                                     default:
-                                        std::cerr << "run.cpp:401:write failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                        std::cerr << "run.cpp:421:write failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                                         throw "what?";
                                 }
                             } else {
@@ -424,7 +444,7 @@ namespace run{
                                 case EINTR:
                                     break;
                                 default:
-                                    std::cerr << "run.cpp:427:poll failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                    std::cerr << "run.cpp:447:poll failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                                     throw "what?";
                             }
                         } else if(pfds[1].revents & POLLIN || pfds[1].revents & POLLHUP){
@@ -438,12 +458,12 @@ namespace run{
                                     case EINTR:
                                         break;
                                     default:
-                                        std::cerr << "run.cpp:441:upstream read failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
+                                        std::cerr << "run.cpp:461:upstream read failed:" << std::make_error_code(std::errc(errno)).message() << std::endl;
                                         throw "what?";
                                 }
                             }
                         } else {
-                            std::cerr << "run.cpp:446:unrecognized events on poll." << std::endl;
+                            std::cerr << "run.cpp:466:unrecognized events on poll." << std::endl;
                             throw "what?";
                         }
                     }while(errno == EINTR || value_size == val.size());
